QuickFix PerfectPerspective.fx performance
Due to compiler poor optimization of a function with for loops assignment to a constant value, which should be evaluated to a constant value only once and applied to all pixels, I added the Macro "PRECISE_CROPPING" which enables the new feature, which currently is very resource heavy.
Hope @crosire will improve the compiler optimization.